requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==2.0.2
pandas==2.2.2
scikit-learn==1.5.2
mlflow==2.16.2
pytest==8.3.2
httpx==0.27.2
Pillow==12.1.1
